+++ background.js
@@ -174,6 +174,12 @@ var advancedDefaultSettings = [{
         "type": TYPE_SMALLTEXT,
         "description": "Affiche la date d'alerte de l'antécédent en orange si la date est dans moins de x mois (6 par défaut, 0 pour désactiver).",
         "default": 6
+    }, {
+        "name": "preAlertVSM",
+        "type": TYPE_SMALLTEXT,
+        "description": "Alerte si le VSM est échu (rouge) ou bientôt échu (orange).",
+        "longDescription": "Affiche la date d'alerte de la VSM en orange si la date est dans moins de x mois (6 par défaut, -1 pour désactiver), puis en rouge une fois l'année dépassée.",
+        "default": 6
     }],
     "sections": [{
         "name": "Médecin Traitant",